requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
groq==0.4.2
python-dotenv==1.0.1
sqlparse==0.4.4
